feat:added test_user_password_hash test function.
