+++ api/tests/test_vote.py
@@ -1,42 +1,6 @@
-import datetime
-import json
-import time
-from flask import jsonify, session
-import mongoengine
-import requests
-from models.comment_vote import Comment_vote
-from models.user import User
-from models.post import Post
-from models.vote import Vote
-import pytest
 from mongoengine import *
-from views.post import (
-    PostList,
-    PostCreate,
-    PostDetail,
-    UserPost,
-)
 from app import app
-import mongomock
-from tests.test_data import *
-
-
-@pytest.fixture()
-def client():
-    disconnect()
-    connect(
-        "mongoenginetest",
-        host="localhost",
-        mongo_client_class=mongomock.MongoClient,
-        uuidRepresentation="standard",
-    )
-    app.config.update(
-        {
-            "TESTING": True,
-        }
-    )
-    with app.test_client() as client:
-        yield client
+from tests.confest import *
 
 
 def test_vote_list_not_logged_in():
